+++ src/main/kotlin/com/github/azurapi/azurapikotlin/internal/exceptions/EquipmentNotFoundException.kt
@@ -1,0 +1,3 @@
+package com.github.azurapi.azurapikotlin.internal.exceptions
+
+class EquipmentNotFoundException(message: String = "") : ApiException(message)